Transferência de classe dependente do Firestore
